autoagi/autoagi_libs/pyproject.toml
[tool.poetry]
name = "autoagi-libs"
version = "0.1.0"
description = "Shared libraries across NextGen AutoAGI"
authors = ["Md Sulaiman <dev.sulaiman@icloud.com>"]
readme = "README.md"
packages = [{ include = "autoagi_libs" }]

[tool.poetry.dependencies]
colorama = "^0.4.6"
expiringdict = "^1.2.2"
google-cloud-logging = "^3.11.3"
pydantic = "^2.10.5"
pydantic-settings = "^2.7.1"
pyjwt = "^2.10.1"
pytest-asyncio = "^0.25.2"
pytest-mock = "^3.14.0"
python = ">=3.10,<4.0"
python-dotenv = "^1.0.1"
supabase = "^2.11.0"

[tool.poetry.group.dev.dependencies]
redis = "^5.2.1"
ruff = "^0.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
extend-select = ["I"]  # sort dependencies
